se hicieron vañidaciones en archivos separados
